pyproject.toml
[project]
name = "tg-budget-planner"
version = "0.1.0"
description = "Telegram Mini App budget planner"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.128.8",
    "uvicorn[standard]==0.39.0",
    "sqlalchemy[asyncio]==2.0.49",
    "asyncpg==0.31.0",
    "alembic==1.16.5",
    "pydantic==2.13.3",
    "pydantic-settings==2.11.0",
    "python-dateutil==2.9.0.post0",
    "httpx==0.28.1",
    "aiogram==3.22.0",
    "apscheduler<4",
    "structlog==25.5.0",
    "aiohttp==3.13.5",
    "pytz",
]

[project.optional-dependencies]
dev = [
    "pytest==8.4.2",
    "pytest-asyncio==1.2.0",
    "httpx==0.28.1",
]

[tool.uv]
dev-dependencies = [
    "pytest==8.4.2",
    "pytest-asyncio==1.2.0",
    "httpx==0.28.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
